Small changes and cleanup

+++ FileMgr/FileManager.cs
@@ -1,9 +1,7 @@
 ﻿using System.Data.SQLite;
 using System.Diagnostics.CodeAnalysis;
-using System.Security.Principal;
-using System.Text.Json.Serialization;
-using Microsoft.Win32;
 using System.Text.Json;
+using Microsoft.Win32;
 
 namespace FileMgr;
 
@@ -11,7 +9,6 @@ internal class ApplicationConfig
 {
     public required string Delimiter { get; set; }
 }
-
 
 /// <summary>
 ///     Public exposed class for file management using tags and file paths.
@@ -29,21 +26,21 @@ public class FileManager
     private readonly DirectoryInfo _filePath;
 
     /// <summary>
+    ///     Stores persistent application configuration.
+    /// </summary>
+    private ApplicationConfig _config;
+
+    /// <summary>
     ///     The database connection for the program.
     /// </summary>
     private Database _database;
-    
+
     /// <summary>
-    /// Stores persistent application configuration.
+    ///     Stores whether the file manager has been initialised.
     /// </summary>
-    private ApplicationConfig _config;
-    
-    /// <summary>
-    /// Stores whether the file manager has been initialised.
-    /// </summary>
-    private bool _initialised = false;
-    
-    
+    private bool _initialised;
+
+
     /* * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * *
      * Miscellaneous Methods
      */
@@ -58,29 +55,31 @@ public class FileManager
     {
         // Set the file path
         _filePath = filePath;
-        
+
         // Check if a .tagster file exists in the directory.
-        if (!File.Exists(_filePath.FullName + "/.tagster"))
-        {
-            throw new MissingFileException("The .tagster file does not exist.", 0);
-        }
-        
+        if (!File.Exists(_filePath.FullName + "/.tagster")) throw new MissingFileException("The .tagster file does not exist.", 0);
+
         // Read the contents as a json object.
         string json = File.ReadAllText(".tagster");
         _config = JsonSerializer.Deserialize<ApplicationConfig>(json) ?? throw new InvalidOperationException("Json error in .tagster file.");
-        
+
         // Do the registry checks
         DoRegistryChecks();
     }
 
+    // Overload for string path
+    public FileManager(string filePath) : this(new DirectoryInfo(filePath))
+    {
+    }
+
     /// <summary>
-    /// Ensures that the program can run by checking the registry for the LongPathsEnabled key.
+    ///     Ensures that the program can run by checking the registry for the LongPathsEnabled key.
     /// </summary>
     private void DoRegistryChecks()
     {
         // Check the state of the HKEY_LOCAL_MACHINE\SOFTWARE\tagster\LongPathsEnabled registry key
         RegistryKey applicationKeys = Registry.CurrentUser.OpenSubKey(@"SOFTWARE\tagster", true) ?? Registry.CurrentUser.CreateSubKey(@"SOFTWARE\tagster");
-        
+
         // Check if the key exists or is not set to 1.
         if (
             applicationKeys.GetValue("LongPathsEnabled") is null ||
@@ -94,20 +93,18 @@ public class FileManager
             // The key exists and is set to 1, so we can continue.
             return;
         }
-        
+
         // Check the state of the HKEY_LOCAL_MACHINE\SYSTEM\CurrentControlSet\Control\FileSystem\LongPathsEnabled registry key as the program will not work without it.
         RegistryKey machineKeys = Registry.LocalMachine.OpenSubKey(@"SYSTEM\CurrentControlSet\Control\FileSystem", true)!; // We can ignore the nullability warning here as we know the key exists.
 
         // Now check if the key exists or is not set to 1.
         if (
-            machineKeys.GetValue("LongPathsEnabled") is null ||
-            (int)(machineKeys.GetValue("LongPathsEnabled") ?? 0) != 1
-        )
-        {
+                machineKeys.GetValue("LongPathsEnabled") is null ||
+                (int)(machineKeys.GetValue("LongPathsEnabled") ?? 0) != 1
+            )
             // The key does not exist, so we need to create it.
-            machineKeys.SetValue("LongPathsEnabled", 1, RegistryValueKind.DWord); 
-        }
-        
+            machineKeys.SetValue("LongPathsEnabled", 1, RegistryValueKind.DWord);
+
         // Now add a registry key for this app to mark this action as done 
         applicationKeys.SetValue("LongPathsEnabled", 1, RegistryValueKind.DWord);
     }
@@ -116,98 +113,106 @@ public class FileManager
     {
         // Check if the database exists.
         if (!File.Exists(_filePath.FullName + "/database.db"))
-        {
             // Throw an exception if the database does not exist.
             throw new UninitialisedDatabaseException("The database does not exist.");
-        }
         // The database exists, so we can connect to it.
         _database = new Database(_filePath.FullName + "/database.db");
         _initialised = true;
     }
 
     /// <summary>
-    /// Called by the frontend to initialise the directory by creating the database and indexing the files.
+    ///     Called by the frontend to initialise the directory by creating the database and indexing the files.
     /// </summary>
     public void InitialiseDirectory(bool fromExistingSources = false)
     {
         _database = Database.InitialiseNew(_filePath);
-        
+
         // Now enumerate through the files in the directory and add them to the database.
         foreach (FileInfo file in _filePath.EnumerateFiles()) _database.AddFile(file, fromExistingSources);
-        
+
         // Set the initialised flag to true.
         _initialised = true;
     }
-    
-    // Overload for string path
-    public FileManager(string filePath) : this(new DirectoryInfo(filePath)) { }
-    
+
     public void Dispose()
     {
-        _database.Dispose();
+        _config = null!;
+        if (_initialised) _database.Dispose();
     }
-    
+
     /* * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * *
      * Database exposed methods
      */
 
     public DbFile AddFile(FileInfo file, bool tagsFromFileName = false)
     {
-        return _database.AddFile(file, tagsFromFileName);
+        if (!_initialised) throw new UninitialisedDatabaseException("The database has not been initialised.");
+        return _database.AddFile(file, tagsFromFileName, _config.Delimiter);
     }
-    
+
     public DbFile AddFile(string path, bool tagsFromFileName = false)
     {
-        return _database.AddFile(path);
+        if (!_initialised) throw new UninitialisedDatabaseException("The database has not been initialised.");
+        return _database.AddFile(path, tagsFromFileName, _config.Delimiter);
     }
-    
+
     public DbFile GetFile(int id)
     {
+        if (!_initialised) throw new UninitialisedDatabaseException("The database has not been initialised.");
         return _database.GetFile(id);
     }
-    
+
     public DbTag AddTag(string tag, string colour = "")
     {
+        if (!_initialised) throw new UninitialisedDatabaseException("The database has not been initialised.");
         return _database.AddTag(tag, colour);
     }
-    
+
     public DbTag GetTag(int id)
     {
+        if (!_initialised) throw new UninitialisedDatabaseException("The database has not been initialised.");
         return _database.GetTag(id)!;
     }
-    
+
     public DbTag GetTag(string tagName)
     {
+        if (!_initialised) throw new UninitialisedDatabaseException("The database has not been initialised.");
         return _database.GetTag(tagName)!;
     }
-    
+
     public List<DbTag?> GetSimilarTags(string tag)
     {
+        if (!_initialised) throw new UninitialisedDatabaseException("The database has not been initialised.");
         return _database.GetSimilarTags(tag);
     }
-    
+
     public void EditTag(int id, string? tag = null, string? colour = null)
     {
+        if (!_initialised) throw new UninitialisedDatabaseException("The database has not been initialised.");
         _database.EditTag(id, tag, colour);
     }
-    
+
     public void AddTagToFile(int fileId, int tagId)
     {
+        if (!_initialised) throw new UninitialisedDatabaseException("The database has not been initialised.");
         _database.AddTagToFile(fileId, tagId);
     }
-    
+
     public void RemoveTagFromFile(int fileId, int tagId)
     {
+        if (!_initialised) throw new UninitialisedDatabaseException("The database has not been initialised.");
         _database.RemoveTagFromFile(fileId, tagId);
     }
-    
+
     public List<DbTag?> GetTagsForFile(int fileId)
     {
+        if (!_initialised) throw new UninitialisedDatabaseException("The database has not been initialised.");
         return _database.GetTagsForFile(fileId);
     }
-    
+
     public List<DbFile?> GetFilesWithTag(int tagId)
     {
+        if (!_initialised) throw new UninitialisedDatabaseException("The database has not been initialised.");
         return _database.GetFilesWithTag(tagId);
     }
 }
@@ -246,7 +251,7 @@ internal class Database
 
         return database;
     }
-    
+
     public void Dispose()
     {
         _connection.Close();
@@ -261,7 +266,8 @@ internal class Database
     /// </summary>
     /// <param name="file">The file to add.</param>
     /// <param name="tagsFromFileName">Whether to extract the tags from the file name.</param>
-    public DbFile AddFile(FileInfo file, bool tagsFromFileName = true)
+    /// <param name="delimiter">The tag delimiter character.</param>
+    public DbFile AddFile(FileInfo file, bool tagsFromFileName = true, string delimiter = "&")
     {
         if (!file.Exists) throw new FileNotFoundException("The file does not exist.", file.FullName);
 
@@ -281,7 +287,7 @@ internal class Database
         string tagGroup = file.FullName.Split(".")[0]; // The tags are stored in the file name before the first period.
 
         // Split the tags into an array at &
-        string[] tags = tagGroup.Split("&");
+        string[] tags = tagGroup.Split(delimiter);
 
         // Add the tags to the database.
         foreach (string tag in tags)
@@ -292,7 +298,7 @@ internal class Database
             // Add the tag to the file.
             AddTagToFile(tagId, dbTag.Id);
         }
-        
+
         return GetFile(tagId);
     }
 
@@ -301,9 +307,10 @@ internal class Database
     /// </summary>
     /// <param name="path">Full path to file to add.</param>
     /// <param name="tagsFromFileName">Whether to extract the tags from the file name.</param>
-    public DbFile AddFile(string path, bool tagsFromFileName = true)
+    /// <param name="delimiter">The tag delimiter character.</param>
+    public DbFile AddFile(string path, bool tagsFromFileName = true, string delimiter = "&")
     {
-        return AddFile(new FileInfo(path), tagsFromFileName);
+        return AddFile(new FileInfo(path), tagsFromFileName, delimiter);
     }
 
     public DbFile GetFile(int id)
@@ -364,7 +371,7 @@ internal class Database
         command.Parameters.AddWithValue("@tag", tag ?? dbTag.Name);
         command.Parameters.AddWithValue("@colour", colour ?? dbTag.Colour);
         command.Parameters.AddWithValue("@id", id);
-        
+
         // Execute the command.
         command.ExecuteNonQuery();
     }
@@ -392,7 +399,7 @@ internal class Database
         // Create query
         SQLiteCommand command = new("SELECT * FROM tags WHERE tag = @tag;", _connection);
         command.Parameters.AddWithValue("@tag", tagName);
-        
+
         // Get the data 
         SQLiteDataReader reader = command.ExecuteReader();
         reader.Read();
@@ -402,10 +409,10 @@ internal class Database
             reader.Close();
             return null;
         }
-        
+
         // Create a new tag object.
         DbTag tag = new(reader.GetInt32(0), reader.GetDateTime(1), reader.GetString(2), reader.GetString(3));
-        
+
         // Close the reader and return the tag.
         reader.Close();
         return tag;
@@ -421,7 +428,7 @@ internal class Database
         // Create query
         SQLiteCommand command = new("SELECT * FROM tags WHERE tag LIKE %@tag%;", _connection);
         command.Parameters.AddWithValue("@tag", tag);
-        
+
         // Get the tags.
         SQLiteDataReader reader = command.ExecuteReader();
         List<DbTag?> tags = [];
@@ -567,4 +574,8 @@ public class MissingFileException(string message, int id) : Exception(message)
 /// <param name="message">Error message</param>
 public class DuplicateTagsException(string message) : Exception(message);
 
+/// <summary>
+///     Used when the database has not been initialised.
+/// </summary>
+/// <param name="message">Error message </param>
 public class UninitialisedDatabaseException(string message) : Exception(message);
